requirements.txt
certifi==2018.4.16
chardet==3.0.4
haishoku-mod==1.1.7
idna==2.6
Pillow==5.1.0
praw==5.4.0
prawcore==0.14.0
pyimgur==0.6.0
requests==2.18.4
update-checker==0.16
urllib3==1.22
